fix: address remaining review feedback on PR #153
Addresses NatSilprasert's requested changes:

1. Routes moved from /dashboard/** to /club/dashboard/**: My Posts at
   /club/dashboard/posts (no more bare index route — /club/dashboard
   itself redirects there), Create Post at
   /club/dashboard/posts/upload (renamed from posts/new), Manage
   Profile at /club/dashboard/profile. Navbar.tsx's
   CLUB_DASHBOARD_NAV_ITEMS/isClubDashboardNavActive and the desktop
   "แดชบอร์ด" CTA updated to match — both were still pointing at the
   old paths after the manual directory move, which would have 404'd.
   isClubDashboardNavActive's "exact match" special case generalized
   from hardcoding "/dashboard" to detecting any item that's a path
   prefix of a sibling (now /club/dashboard/posts, since
   .../posts/upload sits under it).

2. Sidebar now fills the full body height with logout pinned to the
   bottom. Root cause was one level up: (site)/layout.tsx's content
   wrapper was flex-1 (a flex item) but not itself `display: flex`, so
   DashboardShell's `h-full` had nothing definite to resolve a
   percentage against and silently fell back to content-based height.
   Fixed by making that wrapper `flex flex-col` too and having
   DashboardShell/DashboardSidebar grow into it via flex-grow instead
   of a percentage height, which is what actually propagates through a
   flex-grown ancestor. Verified by measuring computed layout
   (aside/main now report identical, full heights) across all three
   dashboard routes, not just visually.

3. Unified the mobile drawer's active-item style: the top nav links
   (ชมรม/กิจกรรม/เกี่ยวกับ) used text-primary only, while the merged
   dashboard section used a filled bg-primary/text-white pill — two
   different "active" treatments in the same drawer. Nav links now use
   the same filled pill as the dashboard section.
